// frontend/src/app/core/generated/simulator.ts
export interface CalculateBNPLResponse {
    /**
     * @generated from protobuf field: int64 cashback_cents = 1
     */
    cashbackCents: bigint;
    /**
     * @generated from protobuf field: int32 cashback_percents = 2
     */
    cashbackPercents: number;
    /**
     * @generated from protobuf field: repeated int64 installments = 3
     */
    installments: bigint[];
}
// ...
class CalculateBNPLResponse$Type extends MessageType<CalculateBNPLResponse> {
    constructor() {
        super("simulator.v1.CalculateBNPLResponse", [
            { no: 1, name: "cashback_cents", kind: "scalar", T: 3 /*ScalarType.INT64*/, L: 0 /*LongType.BIGINT*/ },
            { no: 2, name: "cashback_percents", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 3, name: "installments", kind: "scalar", repeat: 1 /*RepeatType.PACKED*/, T: 3 /*ScalarType.INT64*/, L: 0 /*LongType.BIGINT*/ }
        ]);
    }
    create(value?: PartialMessage<CalculateBNPLResponse>): CalculateBNPLResponse {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.cashbackCents = 0n;
        message.cashbackPercents = 0;
        message.installments = [];
        if (value !== undefined)
            reflectionMergePartial<CalculateBNPLResponse>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: CalculateBNPLResponse): CalculateBNPLResponse {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* int64 cashback_cents */ 1:
                    message.cashbackCents = reader.int64().toBigInt();
                    break;
                case /* int32 cashback_percents */ 2:
                    message.cashbackPercents = reader.int32();
                    break;
                case /* repeated int64 installments */ 3:
                    if (wireType === WireType.LengthDelimited)
                        for (let e = reader.int32() + reader.pos; reader.pos < e;)
                            message.installments.push(reader.int64().toBigInt());
                    else
                        message.installments.push(reader.int64().toBigInt());
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: CalculateBNPLResponse, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* int64 cashback_cents = 1; */
        if (message.cashbackCents !== 0n)
            writer.tag(1, WireType.Varint).int64(message.cashbackCents);
        /* int32 cashback_percents = 2; */
        if (message.cashbackPercents !== 0)
            writer.tag(2, WireType.Varint).int32(message.cashbackPercents);
        /* repeated int64 installments = 3; */
        if (message.installments.length) {
            writer.tag(3, WireType.LengthDelimited).fork();
            for (let i = 0; i < message.installments.length; i++)
                writer.int64(message.installments[i]);
            writer.join();
        }
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message simulator.v1.CalculateBNPLResponse
 */
export const CalculateBNPLResponse = new CalculateBNPLResponse$Type();